+++ public_html/sitemap.php
@@ -40,7 +40,7 @@ $rickies_events__params = [
 	'sort' => $airtable_sorting,
 ];
 include '../includes/data_controllers/event_data_controller.php';
-$home = $latest = $latest_keynote = $latest_annual = $ungraded_filter = $wwdc_filter = $keynote_filter = $annual_filter = $pickies_filter = true;
+$home = $latest = $latest_keynote = $latest_annual = $ungraded_filter = $wwdc_filter = $keynote_filter = $annual_filter = $pickies_filter = $charities = $apple_events = true;
 
 foreach ($rickies_events__array as $event) {
 	if (isset($home)) {
@@ -52,9 +52,17 @@ foreach ($rickies_events__array as $event) {
 		echo sitemap_url('/latest', $event['last_edited'], '0.7');
 		unset($latest);
 	}
+	if (isset($charities)) {
+		echo sitemap_url('/charities', $event['last_edited'], '0.7');
+		unset($charities);
+	}
 	if ($event['type'] == 'keynote' && isset($latest_keynote)) {
 		echo sitemap_url('/latest-keynote', $event['last_edited']);
 		unset($latest_keynote);
+	}
+	if ($event['type'] == 'keynote' && isset($apple_events)) {
+		echo sitemap_url('/apple-events', $event['last_edited']);
+		unset($apple_events);
 	}
 	if ($event['type'] == 'keynote' && isset($keynote_filter)) {
 		echo sitemap_url('/keynote', $event['last_edited'], '0.3');
@@ -157,4 +165,15 @@ echo sitemap_url('/3j-archive', $pick['last_edited'], '0.2', 'monthly');
 // Date for /about
 $about = max([filemtime('../includes/view_controllers/about_controller.php'), filemtime('../includes/about.html')]);
 echo sitemap_url('/about', date('c', $about), '0.5', 'yearly');
+
+// Date for /trophies
+$trophies = max([
+	filemtime('../includes/view_controllers/trophies-controller.php'),
+	filemtime('../includes/views/trophies.php'),
+	filemtime('../includes/trophies-magtricky.php'),
+	filemtime('../includes/trophies-tricky.php'),
+	filemtime('../includes/trophies-ricky.html'),
+	filemtime('../includes/trophies-others.html'),
+]);
+echo sitemap_url('/trophies', date('c', $trophies), '0.5', 'yearly');
 echo '</urlset>';